Move all the things into separate modules
