+++ girderfs/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '1.0'
-__author__ = 'Kacper Kowalik'
+__version__ = "1.0"
+__author__ = "Kacper Kowalik"
