+++ setup.cfg
@@ -1,0 +1,7 @@
+[flake8]
+max-line-length: 100
+show-pep8: False
+show-source: False
+max-complexity: 14
+format: pylint
+ignore: D100,D101,D102,D103,D104,D105,D107,D200,D201,D202,D203,D204,D205,D400,D401,D402,E123,E226,E241,N802,N803,N806,N812,W606,B902
